+++ boilerplates/lucia-auth/files/database/sqliteDb.ts
@@ -1,35 +1,29 @@
 import sqlite, { type Database } from "better-sqlite3";
 
-export const db: Database = sqlite(":memory:");
+export const sqliteDb: Database = sqlite(":memory:");
 
 /**
  * SQLite Schema
  *
  * @link {@see https://lucia-auth.com/database/sqlite#schema}
  */
-db.exec(`CREATE TABLE IF NOT EXISTS users (
+sqliteDb.exec(`CREATE TABLE IF NOT EXISTS users (
     id TEXT NOT NULL PRIMARY KEY,
     username TEXT NOT NULL,
     password_hash TEXT
 )`);
 
-db.exec(`CREATE TABLE IF NOT EXISTS oauth_accounts (
+sqliteDb.exec(`CREATE TABLE IF NOT EXISTS oauth_accounts (
     provider_id TEXT NOT NULL,
     provider_user_id TEXT NOT NULL,
     user_id TEXT NOT NULL,
     PRIMARY KEY (provider_id, provider_user_id),
-    FOREIGN KEY (user_id) REFERENCES users(id)
+    FOREIGN KEY (user_id) REFERENCES users(id) ON UPDATE cascade ON DELETE cascade
 )`);
 
-db.exec(`CREATE TABLE IF NOT EXISTS sessions (
+sqliteDb.exec(`CREATE TABLE IF NOT EXISTS sessions (
     id TEXT NOT NULL PRIMARY KEY,
     expires_at INTEGER NOT NULL,
     user_id TEXT NOT NULL,
-    FOREIGN KEY (user_id) REFERENCES users(id)
+    FOREIGN KEY (user_id) REFERENCES users(id) ON UPDATE cascade ON DELETE cascade
 )`);
-
-export interface DatabaseUser {
-  id: string;
-  username: string;
-  password_hash?: string;
-}
