+++ boilerplates/lucia-auth/files/vike.d.ts
@@ -1,0 +1,11 @@
+import type { User } from "lucia";
+
+declare global {
+  namespace Vike {
+    interface PageContext {
+      user?: User;
+    }
+  }
+}
+
+export {};
